Add --disable-plugins option
By default, images that have plugins (like nova and horizon) will have a
bunch of plugins automatically loaded and installed in to the image.
This increases both the size of the container and the complexity of the
build. In particular, it makes it a bit harder to build custom versions
of images that involve a lot of plugins, as it increases the liklihood
you will run in to a dependency conflict or other issue.

Provide an option to disable plugins on a case-by-case basis as needed.