Allow building images if parent was skipped
This fixes a bug where if an image had a parent that was skipped, it
would never get built. Normally when a parent finishes building it
triggers the build of its children. This process is kicked off by Kolla
finding the grandparents that have no parents above them. However,
sometimes parents should be skipped, and in this case they would never
get scheduled, and so their children would never build, producing a
strange error without much information.

Fix is to allow building an image if its parent is explicitly marked as
having been skipped.