Remove build date from Docker image labels
The temporal tie of this label to a 24-hour period means that any base
images cannot be re-used as a base to build from after the day
increments. While this is likely by design for reproducible builds, it's
a tradeoff that really makes builds slow when they're doing somewhat
infrequently (like every other day or so).

An alternative strategy is to not include the build date, but simply
manually trigger a rebuild of the base images periodically on some sane
schedule (e.g. monthly or when required by a security patch.)

Change-Id: Iedf2c94cda3d31aac3aff1df8298051d80d4a2fa