Fix incorrect parent image status
If using a filter, all of the images are looked at, and the ones that
match the filter are processed. The entire ancestor chain is processed
as part of this step, and images are either marked as unbuildable,
skipped (if in cache or using --skip-parents), or unmatched (the
default.) However, the ancestor images would get processed multiple
times: once during the ancestor loop for a child image, and then
possibly other times for other ancestor loops, and once for the image
itself. The 'else' statement in this loop would always default to
setting the image status to unmatched, potentially overriding a previous
decision to mark it as skipped. This is not correct.

The fix is to only explicitly mark an image as unmatched if it has not
had any status changes (is still in default status=unprocessed state.)

Change-Id: Ibcfdb3c7917221b0cb6744a73512128ff764333d