Do not skip explicitly matched parents
It is sometimes reasonable to have a filter that is generic, like
^nova-, which will build all nova images, including nova-base. However,
if --skip-parents is used in this case, nova-base would not be rebuilt,
because technically it is a parent of some matched image.

The solution here is to only skip the parent if it is not explicitly
matched. Also, give the parent a chance to become matched if it was
marked as skipped; the filter loop will no longer ignore images that
were earlier marked as skipped.

Change-Id: Iff5a075defdb95d393d93cdf51c542fb88835ca0