Add support for specifying parent tag
This tag can be used to re-use a common ancestor tree when trying to
build a feature branch of a descendent part of the tree. Testing out on
Horizon.

Instead of building the parent/child relationship using the canonical
name, which depends on the full image path, link using only the simple
name. This allows having different tags on parents and children, if
desired. The squash functionality is updated to grab the canonical name
from the parent explicitly; if not present, it assumes the base image is
the parent.

Change-Id: Ic6bb8023afafe84dd2e744c9c5c118c1df70e642