Preserve uid/git bits on plugins/* and archives/*
Even though each individual plugin/archive is properly chown'd to
get around the Git CVE issue, when they are processed as a bulk
archive, the .extractall method on the tarfile library will extract
them as the current user and not override this if the user is not
root. It will not log or warn about this.