Handle tarfile extract and overwrite of readonly file
In the case where a tarfile contains reasonly files, and we keep
the same working directory between runs, a second extraction will
fail when it attempts to overwrite the existing file.
This is very common for git pack files, among others.

GNU Tar will prompt in this case, but can proceed, however python
tarfile.extractall does not.

This change simply deletes the offending file from the target and
tries again.